+++ assets/plugins/wwsu-calendar/js/wwsu-calendar-web.js
@@ -105,6 +105,8 @@ class CalendarDb {
 
 		let events = [];
 
+		let scheduleIDs = [];
+
 		/**
 		 * Extends this.processRecord by filtering out events that do not fall within start and end.
 		 *
@@ -136,6 +138,7 @@ class CalendarDb {
 
 		// Define a sort function for schedule types that prioritizes certain types above others in the event of multiple overrides.
 		const scheduleCompare = (a, b) => {
+			/*
 			if (a.scheduleType === "canceled" && b.scheduleType !== "canceled")
 				return -1;
 			if (b.scheduleType === "canceled" && a.scheduleType !== "canceled")
@@ -174,13 +177,14 @@ class CalendarDb {
 				a.scheduleType !== "canceled-changed"
 			)
 				return 1;
+				*/
 			if (a.ID < b.ID) return -1;
 			if (b.ID < a.ID) return 1;
 			return 0;
 		};
 
 		// Define a comparison function that will order calendar events by start time when we run the iteration
-		const compare = function (a, b) {
+		const compare = function(a, b) {
 			try {
 				if (moment(a.start).valueOf() < moment(b.start).valueOf()) {
 					return -1;
@@ -210,7 +214,7 @@ class CalendarDb {
 		};
 
 		/**
-		 * Process a schedule entry.
+		 * Process a (normal) schedule entry.
 		 *
 		 * @param {object} calendar Calendar event the schedule belongs to.
 		 * @param {object} schedule Schedule record.
@@ -228,82 +232,124 @@ class CalendarDb {
 				}
 			}
 
-			// First, process one-time dates/times
-			if (schedule.oneTime && schedule.oneTime.length > 0) {
-				schedule.oneTime.map((oneTime) => {
-					let tempSchedules = [];
-					let scheduleIDs = [];
-					try {
-						// Get schedule overrides if they exist
-						let scheduleOverrides =
-							scheduledb.find(function () {
-								return (
-									this.calendarID === calendar.ID &&
-									this.scheduleID === schedule.ID &&
-									this.scheduleType &&
-									this.scheduleType !== "unscheduled" &&
-									this.originalTime &&
-									moment(this.originalTime).isSame(moment(oneTime), "minute")
-								);
-							}) || [];
-						if (scheduleOverrides.length > 0) {
-							scheduleOverrides.map((exc) => {
-								scheduleIDs.push(exc.ID);
-								tempSchedules.push(exc);
-
-								// For updated records, add a canceled-changed record into the events so people know the original time was changed.
-								if (
-									["updated", "updated-system"].indexOf(exc.scheduleType) !==
-										-1 &&
-									exc.newTime
-								) {
-									_processRecord(
-										tempCal,
-										{
-											calendarID: calendar.ID,
-											ID: schedule.ID,
-											scheduleType: "canceled-changed",
-											scheduleReason: `[SYSTEM] Event was rescheduled to ${moment(
-												exc.newTime
-											).format("llll Z")}`,
-										},
-										exc.originalTime
-									);
-								}
-							});
-						}
-					} catch (e) {
-						console.error(e);
-					}
-
-					// Merge all schedule overrides into one according to scheduleCompare priorities
-					if (tempSchedules.length > 0) {
-						let tempEvent = {};
-						tempSchedules.sort(scheduleCompare).reverse();
-						tempSchedules.map((ts) => {
-							for (let stuff in ts) {
-								if (Object.prototype.hasOwnProperty.call(ts, stuff)) {
-									if (typeof ts[stuff] !== "undefined" && ts[stuff] !== null)
-										tempEvent[stuff] = ts[stuff];
-								}
-							}
-						});
-						_processRecord(tempCal, tempEvent, oneTime);
-					} else {
-						_processRecord(calendar, schedule, oneTime);
-					}
-				});
+			// Get all schedule overrides in memory
+			let scheduleOverrides;
+			try {
+				scheduleOverrides =
+					scheduledb.find(function() {
+						return (
+							this.calendarID === calendar.ID &&
+							this.scheduleType &&
+							this.scheduleType !== "unscheduled"
+						);
+					}) || [];
+			} catch (e) {
+				console.error(e);
 			}
 
-			// Next, process recurring schedules if startTime is not null (we will never process filters if startTime is null)
+			// Preload override processor (date = occurrence date/time to check overrides for)
+			const processOverridesForOccurrence = date => {
+				let overrideProcessed = false;
+				// Get top-level overrides
+				if (scheduleOverrides && scheduleOverrides.length > 0) {
+					scheduleOverrides
+						.filter(
+							override =>
+								override.scheduleID === schedule.ID &&
+								override.originalTime &&
+								moment(override.originalTime).isSame(moment(date), "minute")
+						)
+						.map(override => {
+							// We are going to combine all overrides together into one
+							let tempEvent = override;
+
+							if (
+								["updated", "updated-system"].indexOf(override.scheduleType) !==
+									-1 &&
+								override.newTime
+							) {
+								_processRecord(
+									tempCal,
+									{
+										calendarID: override.calendarID,
+										ID: override.scheduleID,
+										scheduleType: "canceled-changed",
+										scheduleReason: `[SYSTEM] Event was rescheduled to ${moment(
+											override.newTime
+										).format("llll Z")}`
+									},
+									override.originalTime
+								);
+							}
+
+							// Loop through all overrides and merge with tempEvent
+							let currentOverride = override;
+							while (currentOverride) {
+								currentOverride = scheduleOverrides.find(
+									override2 =>
+										override2.calendarID === calendar.ID &&
+										override2.scheduleID === currentOverride.ID
+								);
+								if (currentOverride) {
+									// Merge properties of the override
+									for (let stuff in currentOverride) {
+										if (
+											Object.prototype.hasOwnProperty.call(
+												currentOverride,
+												stuff
+											)
+										) {
+											if (
+												typeof currentOverride[stuff] !== "undefined" &&
+												currentOverride[stuff] !== null
+											)
+												tempEvent[stuff] = currentOverride[stuff];
+										}
+									}
+
+									// For updated records, add a canceled-changed record into the events so people know the original time was changed.
+									if (
+										["updated", "updated-system"].indexOf(
+											currentOverride.scheduleType
+										) !== -1 &&
+										currentOverride.newTime
+									) {
+										_processRecord(
+											tempCal,
+											{
+												calendarID: currentOverride.calendarID,
+												ID: currentOverride.scheduleID,
+												scheduleType: "canceled-changed",
+												scheduleReason: `[SYSTEM] Event was rescheduled to ${moment(
+													currentOverride.newTime
+												).format("llll Z")}`
+											},
+											currentOverride.originalTime
+										);
+									}
+								}
+							}
+
+							// Process the one-time event
+							overrideProcessed = true;
+							_processRecord(tempCal, tempEvent, date);
+						});
+				}
+				if (!overrideProcessed) _processRecord(calendar, schedule, date);
+			};
+
+			// Process one-time dates/times
+			if (schedule.oneTime && schedule.oneTime.length > 0) {
+				schedule.oneTime.map(oneTime => processOverridesForOccurrence(oneTime));
+			}
+
+			// Process recurring schedules if startTime is not null (we will never process filters if startTime is null)
 			if (schedule.startTime && moment(end).isSameOrAfter(moment(start))) {
 				// Construct the moment recurrence
 				let recur = moment.recur({
-					start: start,
-					end: end,
-					rules: schedule.recurrenceRules
-						? schedule.recurrenceRules
-						: undefined,
+					start: schedule.startDate && moment(schedule.startDate).isAfter("2000-01-01") ? schedule.startDate : start,
+					end: schedule.endDate,
+					rules: schedule.recurrenceRules ? schedule.recurrenceRules : undefined
 				});
 
 				// get all the matching dates
@@ -311,7 +357,7 @@ class CalendarDb {
 
 				// loop through all dates
 				if (allDates && allDates.length > 0) {
-					allDates.map((eventStart) => {
+					allDates.map(eventStart => {
 						// If a recurrence interval is specified, skip applicable dates.
 						// NOTE: Combining intervals and calendar rules with moment-recur does not work, so we only use calendar rules for moment-recur.
 						if (
@@ -367,95 +413,15 @@ class CalendarDb {
 							}
 						}
 
-						let tempSchedules = [];
-						let scheduleIDs = [];
-
-						// Get schedule overrides if they exist
-						try {
-							let tempMeta = this.meta; // this.meta scope is not available in scheduledb.find; we must create a temp letiable for it.
-							let scheduleOverrides =
-								scheduledb.find(function () {
-									return (
-										this.calendarID === calendar.ID &&
-										this.scheduleID === schedule.ID &&
-										this.scheduleType &&
-										this.scheduleType !== "unscheduled" &&
-										this.originalTime &&
-										moment(this.originalTime).isSame(
-											moment(
-												`${eventStart}T${schedule.startTime}${moment(eventStart)
-													.tz(
-														tempMeta
-															? tempMeta.meta.timezone
-															: moment.tz.guess()
-													)
-													.format("Z")}`
-											),
-											"minute"
-										)
-									);
-								}) || [];
-							if (scheduleOverrides.length > 0) {
-								scheduleOverrides.map((exc) => {
-									scheduleIDs.push(exc.ID);
-									tempSchedules.push(exc);
-
-									// For updated records, add a canceled-changed record into the events so people know the original time was changed.
-									if (
-										["updated", "updated-system"].indexOf(exc.scheduleType) !==
-											-1 &&
-										exc.newTime
-									) {
-										_processRecord(
-											calendar,
-											{
-												calendarID: calendar.ID,
-												ID: schedule.ID,
-												scheduleType: "canceled-changed",
-												scheduleReason: `[SYSTEM] Event was rescheduled to ${moment
-													.parseZone(exc.newTime)
-													.format("llll Z")}`,
-											},
-											exc.originalTime
-										);
-									}
-								});
-							}
-						} catch (e) {
-							console.error(e);
-						}
-
-						// Merge all schedule overrides into one according to scheduleCompare priorities
-						if (tempSchedules.length > 0) {
-							let tempEvent = {};
-							tempSchedules.sort(scheduleCompare).reverse();
-							tempSchedules.map((ts) => {
-								for (let stuff in ts) {
-									if (Object.prototype.hasOwnProperty.call(ts, stuff)) {
-										if (typeof ts[stuff] !== "undefined" && ts[stuff] !== null)
-											tempEvent[stuff] = ts[stuff];
-									}
-								}
-							});
-							_processRecord(
-								tempCal,
-								tempEvent,
-								`${eventStart}T${schedule.startTime}${moment(eventStart)
-									.tz(this.meta ? this.meta.meta.timezone : moment.tz.guess())
-									.format("Z")}`
-							);
-						} else {
-							_processRecord(
-								calendar,
-								schedule,
-								`${eventStart}T${schedule.startTime}${moment(eventStart)
-									.tz(this.meta ? this.meta.meta.timezone : moment.tz.guess())
-									.format("Z")}`
-							);
-						}
+						processOverridesForOccurrence(
+							`${eventStart}T${schedule.startTime}${moment(eventStart)
+								.tz(this.meta ? this.meta.meta.timezone : moment.tz.guess())
+								.format("Z")}`
+						);
 					});
 				}
 			}
+
 			taskComplete();
 		};
 
@@ -464,29 +430,37 @@ class CalendarDb {
 		 *
 		 * @param {object} calendar The calendar record.
 		 */
-		const processCalendarEntry = (calendar) => {
-			// Get regular and unscheduled events
+		const processCalendarEntry = calendar => {
+			// Get all schedules for the provided calendar
 			let regularEvents = scheduledb.find({
-				calendarID: calendar.ID,
-				scheduleType: [null, "unscheduled", undefined],
+				calendarID: calendar.ID
 			});
-			regularEvents.map((schedule) => {
-				// Add to task queue
-				tasks++;
-				if (callback) {
-					this.queue.add(() => {
+
+			// Get regular and unscheduled events
+			regularEvents
+				.filter(
+					schedule =>
+						[null, "unscheduled", undefined].indexOf(schedule.scheduleType) !==
+						-1
+				)
+				.sort(scheduleCompare)
+				.map(schedule => {
+					// Add to task queue
+					tasks++;
+					if (callback) {
+						this.queue.add(() => {
+							processScheduleEntry(calendar, schedule);
+						});
+					} else {
 						processScheduleEntry(calendar, schedule);
-					});
-				} else {
-					processScheduleEntry(calendar, schedule);
-				}
-			});
+					}
+				});
 			taskComplete();
 		};
 
 		// Get all calendar events and process their schedules
 		let results = calendardb.find(query);
-		results.map((calendar) => {
+		results.map(calendar => {
 			// Add to task queue
 			tasks++;
 			if (callback) {
@@ -520,10 +494,10 @@ class CalendarDb {
 		 *
 		 * @param {array} events Array of events
 		 */
-		const afterFunction = (events) => {
+		const afterFunction = events => {
 			if (events.length > 0) {
 				// Order events by priority (priority value, then start time, then ID)
-				const compare = function (a, b) {
+				const compare = function(a, b) {
 					try {
 						if (a.priority > b.priority) {
 							return -1;
@@ -553,7 +527,7 @@ class CalendarDb {
 				let returnData = [];
 
 				events
-					.filter((event) => {
+					.filter(event => {
 						// Canceled events should not be playing
 						if (
 							event.scheduleType === "canceled" ||
@@ -600,7 +574,7 @@ class CalendarDb {
 							);
 						}
 					})
-					.map((event) => {
+					.map(event => {
 						if (event && event.unique) returnData.push(event);
 					});
 
@@ -765,7 +739,7 @@ class CalendarDb {
 				overriddenID: overrides.scheduleID || null,
 				type: overridden.type,
 				hosts: overridden.hosts,
-				name: overridden.name,
+				name: overridden.name
 			};
 
 			let startdiff = moment(overrides.start).diff(
@@ -811,12 +785,11 @@ class CalendarDb {
 				);
 			} else {
 				let duplicate = additions.filter(
-					(rec) =>
+					rec =>
 						rec.scheduleID === newRecord.scheduleID &&
 						rec.scheduleType === newRecord.scheduleType &&
 						rec.originalTime === newRecord.originalTime
 				);
-				console.log(duplicate.length);
 				if (duplicate.length < 1) additions.push(newRecord);
 			}
 		};
@@ -826,7 +799,7 @@ class CalendarDb {
 		 *
 		 * @param {function} cb Callback fired when all tasks in this stage are complete.
 		 */
-		const taskComplete3 = (cb) => {
+		const taskComplete3 = cb => {
 			tasksCompleted++;
 			let newprogress = tasksCompleted > 0 ? tasksCompleted / tasks : 0;
 			progressCallback(
@@ -844,7 +817,7 @@ class CalendarDb {
 		 *
 		 * @param {function} cb Callback fired when all tasks in this stage are complete.
 		 */
-		const taskComplete = (cb) => {
+		const taskComplete = cb => {
 			tasksCompleted++;
 			let newprogress = tasksCompleted > 0 ? tasksCompleted / tasks : 0;
 			progressCallback(
@@ -876,7 +849,7 @@ class CalendarDb {
 		 *
 		 * @param {object} query Query to process
 		 */
-		const processQuery = (query) => {
+		const processQuery = query => {
 			// Run the query in our copied schedule db
 			if (typeof query.remove !== "undefined") {
 				query.remove = vschedule.find({ ID: query.remove }, true);
@@ -918,7 +891,7 @@ class CalendarDb {
 							end: moment(query[key].originalTime)
 								.startOf("minute")
 								.add(event.duration, "minutes")
-								.toISOString(true),
+								.toISOString(true)
 						});
 					}
 					if (query[key].newTime) {
@@ -932,7 +905,9 @@ class CalendarDb {
 						}
 						if (
 							!end ||
-							moment(query[key].newTime).startOf("minute").isAfter(moment(end))
+							moment(query[key].newTime)
+								.startOf("minute")
+								.isAfter(moment(end))
 						) {
 							end = moment(query[key].newTime)
 								.startOf("minute")
@@ -945,28 +920,37 @@ class CalendarDb {
 							end: moment(query[key].newTime)
 								.startOf("minute")
 								.add(event.duration, "minutes")
-								.toISOString(true),
+								.toISOString(true)
 						});
 					}
 					if (query[key].oneTime && query[key].oneTime.length > 0) {
-						query[key].oneTime.map((ot) => {
+						query[key].oneTime.map(ot => {
 							if (
 								!start ||
-								moment(ot).startOf("minute").isBefore(moment(start))
+								moment(ot)
+									.startOf("minute")
+									.isBefore(moment(start))
 							) {
 								start = moment(ot).startOf("minute");
 							}
-							if (!end || moment(ot).startOf("minute").isAfter(moment(end))) {
+							if (
+								!end ||
+								moment(ot)
+									.startOf("minute")
+									.isAfter(moment(end))
+							) {
 								end = moment(ot)
 									.startOf("minute")
 									.add(event.duration, "minutes");
 							}
 							timePeriods.push({
-								start: moment(ot).startOf("minute").toISOString(true),
+								start: moment(ot)
+									.startOf("minute")
+									.toISOString(true),
 								end: moment(ot)
 									.startOf("minute")
 									.add(event.duration, "minutes")
-									.toISOString(true),
+									.toISOString(true)
 							});
 						});
 					}
@@ -977,13 +961,17 @@ class CalendarDb {
 					) {
 						if (
 							!start ||
-							moment(event.startDate).startOf("minute").isBefore(moment(start))
+							moment(event.startDate)
+								.startOf("minute")
+								.isBefore(moment(start))
 						) {
 							start = moment(event.startDate).startOf("minute");
 						}
 						if (
 							!end ||
-							moment(event.endDate).startOf("minute").isAfter(moment(end))
+							moment(event.endDate)
+								.startOf("minute")
+								.isAfter(moment(end))
 						) {
 							end = moment(event.endDate)
 								.startOf("minute")
@@ -997,9 +985,7 @@ class CalendarDb {
 							let recur = moment.recur({
 								start: start,
 								end: end,
-								rules: event.recurrenceRules
-									? event.recurrenceRules
-									: undefined,
+								rules: event.recurrenceRules ? event.recurrenceRules : undefined
 							});
 
 							// get all the matching dates
@@ -1007,7 +993,7 @@ class CalendarDb {
 
 							// Loop through each schedule between start and end
 							if (allDates && allDates.length > 0) {
-								allDates.map((eventStart) => {
+								allDates.map(eventStart => {
 									// Skip dates that fail recurrence intervals
 									if (
 										event.recurrenceInterval &&
@@ -1077,7 +1063,7 @@ class CalendarDb {
 												this.meta ? this.meta.meta.timezone : moment.tz.guess()
 											)
 											.add(event.duration, "minutes")
-											.toISOString(true),
+											.toISOString(true)
 									});
 								});
 							}
@@ -1099,7 +1085,7 @@ class CalendarDb {
 			if (event.overriddenID) {
 				// Find the original event via the unfiltered events
 				let record = unfilteredEvents.find(
-					(eventb) => eventb.scheduleID === event.overriddenID
+					eventb => eventb.scheduleID === event.overriddenID
 				);
 
 				// If we could not find it, the override is invalid, so we can remove it and not continue beyond this point for the event.
@@ -1113,7 +1099,7 @@ class CalendarDb {
 			// Iterate conflict checking on every event after the index
 			events
 				.filter((ev, ind) => ind > index)
-				.map((ev) => {
+				.map(ev => {
 					try {
 						checkAndResolveConflicts(event, ev);
 					} catch (e) {
@@ -1128,7 +1114,7 @@ class CalendarDb {
 			// Process updateCalendar or removeCalendar before we continue with anything else
 			queries
 				.filter(
-					(query) =>
+					query =>
 						typeof query.updateCalendar !== "undefined" ||
 						typeof query.removeCalendar !== "undefined"
 				)
@@ -1140,10 +1126,10 @@ class CalendarDb {
 						// Now, we need to remove updateCalendar from the query and replace it with all of its schedules as update queries.
 						// That way, we can check all of its schedules for changes in conflicts resulting from changes in calendar defaults.
 						let schedules = vschedule.find({
-							calendarID: query.updateCalendar.ID,
+							calendarID: query.updateCalendar.ID
 						});
 						queries.splice(index, 1);
-						schedules.map((schedule) => {
+						schedules.map(schedule => {
 							queries.push({ update: schedule });
 						});
 					}
@@ -1153,10 +1139,10 @@ class CalendarDb {
 						// Remove the original removeCalendar query as we do not want to process it beyond this map.
 						// We need to add all of the calendar's schedule records as remove queries since they will get removed too.
 						let schedules = vschedule.find({
-							calendarID: query.removeCalendar,
+							calendarID: query.removeCalendar
 						});
 						queries.splice(index, 1);
-						schedules.map((schedule) => {
+						schedules.map(schedule => {
 							queries.push({ remove: schedule.ID });
 						});
 					}
@@ -1172,7 +1158,7 @@ class CalendarDb {
 				}
 			}
 
-			const eventsCall2 = (events) => {
+			const eventsCall2 = events => {
 				// Now, go through every event for conflict checking
 				tasks = events.length;
 				tasksCompleted = 0;
@@ -1203,10 +1189,9 @@ class CalendarDb {
 			 *
 			 * @param {array} events Array of events
 			 */
-			const eventsCall = (events) => {
+			const eventsCall = events => {
 				progressCallback(`Stage 3 of 4: Intelligently filtering events`);
 				unfilteredEvents = _.cloneDeep(events); // Set unfiltered events to the variable; used for some conflict checks
-				console.dir(unfilteredEvents);
 				tasks = events.length;
 				tasksCompleted = 0;
 				let filteredEvents = [];
@@ -1221,12 +1206,12 @@ class CalendarDb {
 					}
 				}
 
-				events.map((event) => {
+				events.map(event => {
 					// Called on each event to determine of its start/end times fall within any of the query times.
 					// This speeds up conflict checking by not checking events outside of the dates/times affected by the queries (unless the event overrides another event within the time frame; those are included too).
-					const _determineFilter = (_event) => {
+					const _determineFilter = _event => {
 						let filter = timePeriods.find(
-							(period) =>
+							period =>
 								moment(_event.end).isAfter(moment(period.start)) &&
 								moment(_event.start).isSameOrBefore(moment(period.end))
 						);
@@ -1302,7 +1287,7 @@ class CalendarDb {
 			// Process virtual queries
 			tasks = queries.length;
 			tasksCompleted = 0;
-			queries.forEach((query) => {
+			queries.forEach(query => {
 				if (callback) {
 					this.queue.add(() => {
 						processQuery(query);
@@ -1323,8 +1308,8 @@ class CalendarDb {
 					removals: [],
 					additions: [],
 					errors: [
-						"You must provide at least one query to do conflict checking",
-					],
+						"You must provide at least one query to do conflict checking"
+					]
 				});
 				return;
 			} else {
@@ -1332,8 +1317,8 @@ class CalendarDb {
 					removals: [],
 					additions: [],
 					errors: [
-						"You must provide at least one query to do conflict checking",
-					],
+						"You must provide at least one query to do conflict checking"
+					]
 				};
 			}
 		}
@@ -1352,10 +1337,10 @@ class CalendarDb {
 	 */
 	whoShouldBeIn(callback = null, progressCallback = () => {}) {
 		// Function called after running this.getEvents
-		const afterFunction = (events) => {
+		const afterFunction = events => {
 			if (events.length > 0) {
 				// Sort by start time
-				const compare = function (a, b) {
+				const compare = function(a, b) {
 					try {
 						if (moment(a.start).valueOf() < moment(b.start).valueOf()) {
 							return -1;
@@ -1376,7 +1361,7 @@ class CalendarDb {
 				};
 				events = events.sort(compare);
 
-				events = events.filter((event) => {
+				events = events.filter(event => {
 					if (
 						event.scheduleType === "canceled" ||
 						event.scheduleType === "canceled-system" ||
@@ -1784,11 +1769,20 @@ class CalendarDb {
 			logo: schedule.logo ? schedule.logo : calendar.logo || null, // URL to the event logo
 			banner: schedule.banner ? schedule.banner : calendar.banner || null, // URL to the event banner
 			newTime: schedule.newTime
-				? moment.parseZone(schedule.newTime).startOf("minute").toISOString(true)
+				? moment
+						.parseZone(schedule.newTime)
+						.startOf("minute")
+						.toISOString(true)
 				: null, // If an exception is applied that overrides an event's start time, this is the event's new start time.
 			start: schedule.newTime
-				? moment.parseZone(schedule.newTime).startOf("minute").toISOString(true)
-				: moment.parseZone(eventStart).startOf("minute").toISOString(true), // Start time of the event
+				? moment
+						.parseZone(schedule.newTime)
+						.startOf("minute")
+						.toISOString(true)
+				: moment
+						.parseZone(eventStart)
+						.startOf("minute")
+						.toISOString(true), // Start time of the event
 			duration:
 				schedule.duration || schedule.duration === 0
 					? schedule.duration
@@ -1845,7 +1839,7 @@ class CalendarDb {
 			timeChanged:
 				schedule.scheduleID && (schedule.newTime || schedule.duration), // True if this event's time was changed from the original, else false
 			createdAt: schedule.createdAt || calendar.createdAt, // createdAt used to determine which event gets priority in conflict checking if both have the same priority
-			updatedAt: schedule.updatedAt || calendar.updatedAt,
+			updatedAt: schedule.updatedAt || calendar.updatedAt
 		};
 
 		// Determine event color
@@ -1882,7 +1876,10 @@ class CalendarDb {
 						.parseZone(criteria.start)
 						.add(schedule.duration || calendar.duration, "minutes")
 						.toISOString(true)
-				: moment.parseZone(criteria.start).startOf("minute").toISOString(true);
+				: moment
+						.parseZone(criteria.start)
+						.startOf("minute")
+						.toISOString(true);
 
 		return criteria;
 	}
@@ -1895,7 +1892,9 @@ class CalendarDb {
 	 */
 	weekOfMonth(input) {
 		const firstDayOfMonth = moment(input).startOf("month");
-		const firstDayOfWeek = moment(firstDayOfMonth).clone().startOf("week");
+		const firstDayOfWeek = moment(firstDayOfMonth)
+			.clone()
+			.startOf("week");
 
 		const offset = firstDayOfMonth.diff(firstDayOfWeek, "days");
 
@@ -1924,7 +1923,7 @@ class CalendarDb {
 
 		// Add oneTime dates/times to the oneTime letiable.
 		if (event.oneTime && event.oneTime.length > 0) {
-			oneTime = event.oneTime.map((onetime) =>
+			oneTime = event.oneTime.map(onetime =>
 				moment.parseZone(onetime).format("LLLL Z")
 			);
 		}
@@ -1945,7 +1944,7 @@ class CalendarDb {
 				recurDayString += `... and `;
 			}
 
-			event.recurrenceRules.map((rule) => {
+			event.recurrenceRules.map(rule => {
 				let days;
 				if (!rule.measure || !rule.units || rule.units.length === 0) return;
 				switch (rule.measure) {
@@ -1960,7 +1959,7 @@ class CalendarDb {
 					case "monthsOfYear":
 						days = rule.units
 							.sort((a, b) => a - b)
-							.map((unit) => {
+							.map(unit => {
 								switch (unit) {
 									case 0:
 										return "January";
@@ -1994,7 +1993,7 @@ class CalendarDb {
 					case "daysOfWeek":
 						days = rule.units
 							.sort((a, b) => a - b)
-							.map((unit) => {
+							.map(unit => {
 								switch (unit) {
 									case 0:
 										return "Sunday";
@@ -2019,7 +2018,7 @@ class CalendarDb {
 					case "weeksOfMonthByDay":
 						days = rule.units
 							.sort((a, b) => a - b)
-							.map((unit) => {
+							.map(unit => {
 								switch (unit) {
 									case 0:
 										return "1st";
@@ -2042,7 +2041,7 @@ class CalendarDb {
 					case "daysOfMonth":
 						days = rule.units
 							.sort((a, b) => a - b)
-							.map((unit) => {
+							.map(unit => {
 								switch (unit) {
 									case 1:
 									case 21:
@@ -2062,7 +2061,7 @@ class CalendarDb {
 					case "weeksOfYear":
 						days = rule.units
 							.sort((a, b) => a - b)
-							.map((unit) => {
+							.map(unit => {
 								switch (unit) {
 									case 1:
 									case 21:
@@ -2091,7 +2090,11 @@ class CalendarDb {
 			});
 
 			recurDayString += `... at ${event.startTime}${
-				this.meta ? moment().tz(this.meta.meta.timezone).format(" z") : ``
+				this.meta
+					? moment()
+							.tz(this.meta.meta.timezone)
+							.format(" z")
+					: ``
 			}`;
 		}
 
@@ -2117,7 +2120,7 @@ class CalendarDb {
 	}
 
 	/**
-	 * Polyfill missing information in a schedule record from its scheduleID (if applicable) and the calendar event's default properties.
+	 * Polyfill missing information in a schedule record; start from the original calendar ID and work through each schedule sequentially up to this one.
 	 *
 	 * @param {object} _record The schedule database record
 	 * @param {WWSUdb} calendardb If provided, will use this database of calendar events instead of the CalendarDb one.
@@ -2132,23 +2135,34 @@ class CalendarDb {
 		let tempCal = {};
 		let event;
 		let schedule;
+		let schedules = [];
 		let record = _.cloneDeep(_record); // Clone the record to avoid accidental mutable object editing.
 		if (record.calendarID) {
 			let calendar = calendardb.find({ ID: record.calendarID }, true);
 			tempCal = calendar || {};
+
 			if (record.scheduleID) {
 				schedule = scheduledb.find({ ID: record.scheduleID }, true);
-			}
-			if (schedule) {
-				for (let stuff in schedule) {
-					if (Object.prototype.hasOwnProperty.call(schedule, stuff)) {
-						if (
-							typeof schedule[stuff] !== "undefined" &&
-							schedule[stuff] !== null
-						)
-							tempCal[stuff] = schedule[stuff];
-					}
+
+				// Recurse back all schedules
+				while (schedule) {
+					schedules.push(schedule);
+					schedule = scheduledb.find({ ID: schedule.scheduleID }, true);
 				}
+				schedules.reverse(); // Put the schedules in chronological order
+			}
+			if (schedules.length > 0) {
+				schedules.map(schedule => {
+					for (let stuff in schedule) {
+						if (Object.prototype.hasOwnProperty.call(schedule, stuff)) {
+							if (
+								typeof schedule[stuff] !== "undefined" &&
+								schedule[stuff] !== null
+							)
+								tempCal[stuff] = schedule[stuff];
+						}
+					}
+				});
 				event = this.processRecord(
 					tempCal,
 					record,